+++ lib/templates/twig/TwigEngine.php
@@ -1,14 +1,13 @@
 <?php
-
-namespace humhub\modules\custom_pages\lib\templates\twig;
 /**
  * @link https://www.humhub.org/
  * @copyright Copyright (c) 2015 HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
+namespace humhub\modules\custom_pages\lib\templates\twig;
 
-use \humhub\modules\custom_pages\lib\templates\TemplateEngine;
+use humhub\modules\custom_pages\lib\templates\TemplateEngine;
 use Twig\Environment;
 
 /**
@@ -20,9 +19,9 @@ use Twig\Environment;
 class TwigEngine implements TemplateEngine
 {
     /**
-     * @inerhitdoc
+     * @inheritdoc
      * 
-     * @param type $template template name
+     * @param string $template template name
      * @param array $content array input [elementName => content]
      * @return string
      */
@@ -30,7 +29,8 @@ class TwigEngine implements TemplateEngine
     {
         $loader = new DatabaseTwigLoader();
         $twig = new Environment($loader, ['autoescape' => false, 'debug' => true]);
-        return $twig->render($template ,$content);
+
+        return $twig->render($template, $content);
     }
 
 }