+++ models/Page.php
@@ -3,6 +3,7 @@
 namespace humhub\modules\custom_pages\models;
 
 use humhub\modules\content\models\Content;
+use humhub\modules\custom_pages\helpers\Html;
 use humhub\modules\custom_pages\helpers\Url;
 use humhub\modules\custom_pages\models\forms\SettingsForm;
 use humhub\modules\custom_pages\modules\template\models\Template;
@@ -186,6 +187,10 @@ class Page extends CustomContentContainer
      */
     public function getPageContent()
     {
+        if ($this->type == HtmlType::ID) {
+            return Html::applyScriptNonce($this->page_content);
+        }
+
         return $this->page_content;
     }